platform/sbin/apps/rabbitmq_workers/workflow/worker/requirements.txt
pika==1.2.0
simplejson==3.2.0
PyYAML==5.4.1
Jinja2==3.0.2
pandas==1.1.5
